Bug fix.
The way `execute.py` was written, if two methods with the same name but
different methargs were used, the results would all be overwritten in a
single dictionary entry. Note that jug would still run everything
because `jug-execute.py` is written such that any subdirectory with a
`config.yaml` file is executed.

The simple modification here is that the result dictionary now uses the
subdirectory name, i.e., config file key for that particular method,
rather than the method name itself. It is therefore the responsibility
of the user to define a name or label for the particular method
configuration. This is not the most automatic way to do this but I think
it is the most expressive and later this label can also be used in
plots.