Change to replicate making it easier to delete and rerun.
With this change, one can simply delete a method's subdirectory and
rerun the `replicate.py` script on the experiment directory. The
script will recreate clean subdirectories in place of the ones that
were deleted. For all other subdirectories, the script will run
until the seed subdirectory, where existence is checked.

Note however that this is also a good way to increase the number of
repetitions of an experiment while keeping all existing, executed
runs untouched.

Finally, note that we could have left `shutil.copytree()` throw an
exception which would be caught by the try-except clause but this
would cause problems if the behaviour of `shutil.copytree()`
changed at any point in the future.